+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.5 (unreleased)
 
-- Added scanning for sets with Redis
+- Added scanning for lists and sets with Redis
 
 ## 0.1.4 (2022-09-07)